requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
dnspython==2.2.1
idna==3.4
PyYAML==6.0
requests==2.28.1
Sublist3r==1.0
urllib3==1.26.12
